refactor(sourcing): move sourcing to pop-common
