+++ crates/pop-cli/src/commands/up/parachain.rs
@@ -8,7 +8,8 @@ use cliclack::{
 };
 use console::{Emoji, Style, Term};
 use duct::cmd;
-use pop_parachains::{Error, IndexSet, NetworkNode, Status, Zombienet};
+use pop_common::sourcing::binary::Status;
+use pop_parachains::{Error, IndexSet, NetworkNode, Zombienet};
 use std::{path::PathBuf, time::Duration};
 use tokio::time::sleep;
