+++ crates/pop-common/src/lib.rs
@@ -3,6 +3,7 @@ pub mod errors;
 pub mod git;
 pub mod helpers;
 pub mod manifest;
+pub mod sourcing;
 pub mod templates;
 
 pub use build::Profile;
