+++ crates/pop-contracts/src/utils/contracts_node.rs
@@ -1,7 +1,10 @@
 use crate::errors::Error;
 use contract_extrinsics::{RawParams, RpcRequest};
 use duct::cmd;
-use pop_parachains::{Binary, GitHubSource, Source};
+use pop_common::sourcing::{
+	binary::Binary,
+	sourcing::{GitHub as GitHubSource, Source},
+};
 use std::{
 	env::consts::OS,
 	fs::File,
