+++ crates/pop-parachains/src/up/mod.rs
@@ -3,9 +3,13 @@
 use crate::errors::Error;
 use glob::glob;
 use indexmap::IndexMap;
-pub use pop_common::git::GitHub;
-use sourcing::GitHub::*;
-pub use sourcing::{GitHub as GitHubSource, Source, Source::*};
+pub use pop_common::{
+	git::GitHub,
+	sourcing::{
+		binary::{Binary, Repository},
+		sourcing::{GitHub::*, Source, Source::*},
+	},
+};
 use std::{
 	fmt::Debug,
 	fs::write,
@@ -15,14 +19,12 @@ use std::{
 use symlink::{remove_symlink_file, symlink_file};
 use tempfile::{Builder, NamedTempFile};
 use toml_edit::{value, ArrayOfTables, DocumentMut, Formatted, Item, Table, Value};
-pub use url::Url;
 use zombienet_sdk::{Network, NetworkConfig, NetworkConfigExt};
 use zombienet_support::fs::local::LocalFileSystem;
 
 mod chain_specs;
 mod parachains;
 mod relay;
-mod sourcing;
 
 /// Configuration to launch a local network.
 pub struct Zombienet {
@@ -580,224 +582,6 @@ impl Parachain {
 	}
 }
 
-/// A binary used to launch a node.
-#[derive(Debug, PartialEq)]
-pub enum Binary {
-	/// A local binary.
-	Local {
-		/// The name of the binary.
-		name: String,
-		/// The path of the binary.
-		path: PathBuf,
-		/// If applicable, the path to a manifest used to build the binary if missing.
-		manifest: Option<PathBuf>,
-	},
-	/// A binary which needs to be sourced.
-	Source {
-		/// The name of the binary.
-		name: String,
-		/// The source of the binary.
-		#[allow(private_interfaces)]
-		source: Source,
-		/// The cache to be used to store the binary.
-		cache: PathBuf,
-	},
-}
-
-impl Binary {
-	/// Whether the binary exists.
-	pub fn exists(&self) -> bool {
-		self.path().exists()
-	}
-
-	/// If applicable, the latest version available.
-	pub fn latest(&self) -> Option<&str> {
-		match self {
-			Self::Local { .. } => None,
-			Self::Source { source, .. } => {
-				if let GitHub(ReleaseArchive { latest, .. }) = source {
-					latest.as_deref()
-				} else {
-					None
-				}
-			},
-		}
-	}
-
-	/// Whether the binary is defined locally.
-	pub fn local(&self) -> bool {
-		matches!(self, Self::Local { .. })
-	}
-
-	/// The name of the binary.
-	pub fn name(&self) -> &str {
-		match self {
-			Self::Local { name, .. } => name,
-			Self::Source { name, .. } => name,
-		}
-	}
-
-	/// The path of the binary.
-	pub fn path(&self) -> PathBuf {
-		match self {
-			Self::Local { path, .. } => path.to_path_buf(),
-			Self::Source { name, source, cache, .. } => {
-				// Determine whether a specific version is specified
-				let version = match source {
-					Git { reference, .. } => reference.as_ref(),
-					GitHub(source) => match source {
-						ReleaseArchive { tag, .. } => tag.as_ref(),
-						SourceCodeArchive { reference, .. } => reference.as_ref(),
-					},
-					Archive { .. } | Source::Url { .. } => None,
-				};
-				version.map_or_else(|| cache.join(name), |v| cache.join(format!("{name}-{v}")))
-			},
-		}
-	}
-
-	/// Attempts to resolve a version of a binary based on whether one is specified, an existing version
-	/// can be found cached locally, or uses the latest version.
-	///
-	/// # Arguments
-	/// * `name` - The name of the binary.
-	/// * `specified` - If available, a version explicitly specified.
-	/// * `available` - The available versions, used to check for those cached locally or the latest otherwise.
-	/// * `cache` - The location used for caching binaries.
-	fn resolve_version(
-		name: &str,
-		specified: Option<&str>,
-		available: &[impl AsRef<str>],
-		cache: &Path,
-	) -> Option<String> {
-		match specified {
-			Some(version) => Some(version.to_string()),
-			None => available
-				.iter()
-				.map(|v| v.as_ref())
-				// Default to latest version available locally
-				.filter_map(|version| {
-					let path = cache.join(format!("{name}-{version}"));
-					path.exists().then_some(Some(version.to_string()))
-				})
-				.nth(0)
-				.unwrap_or(
-					// Default to latest version
-					available.get(0).and_then(|version| Some(version.as_ref().to_string())),
-				),
-		}
-	}
-
-	/// Sources the binary.
-	///
-	/// # Arguments
-	/// * `release` - Whether any binaries needing to be built should be done so using the release profile.
-	/// * `status` - Used to observe status updates.
-	/// * `verbose` - Whether verbose output is required.
-	pub async fn source(
-		&self,
-		release: bool,
-		status: &impl Status,
-		verbose: bool,
-	) -> Result<(), Error> {
-		match self {
-			Self::Local { name, path, manifest, .. } => match manifest {
-				None => {
-					return Err(Error::MissingBinary(format!(
-						"The {path:?} binary cannot be sourced automatically."
-					)))
-				},
-				Some(manifest) => {
-					sourcing::from_local_package(manifest, name, release, status, verbose).await
-				},
-			},
-			Self::Source { source, cache, .. } => {
-				source.source(cache, release, status, verbose).await
-			},
-		}
-	}
-
-	/// Whether any locally cached version can be replaced with a newer version.
-	pub fn stale(&self) -> bool {
-		// Only binaries sourced from GitHub release archives can currently be determined as stale
-		let Self::Source { source: GitHub(ReleaseArchive { tag, latest, .. }), .. } = self else {
-			return false;
-		};
-		latest.as_ref().map_or(false, |l| tag.as_ref() != Some(l))
-	}
-
-	/// Specifies that the latest available versions are to be used (where possible).
-	pub fn use_latest(&mut self) {
-		if let Self::Source { source: GitHub(ReleaseArchive { tag, latest, .. }), .. } = self {
-			if let Some(latest) = latest {
-				*tag = Some(latest.clone())
-			}
-		};
-	}
-
-	/// If applicable, the version of the binary.
-	pub fn version(&self) -> Option<&str> {
-		match self {
-			Self::Local { .. } => None,
-			Self::Source { source, .. } => match source {
-				Git { reference, .. } => reference.as_ref(),
-				GitHub(source) => match source {
-					ReleaseArchive { tag, .. } => tag.as_ref(),
-					SourceCodeArchive { reference, .. } => reference.as_ref(),
-				},
-				Archive { .. } | Source::Url { .. } => None,
-			},
-		}
-		.map(|r| r.as_str())
-	}
-}
-
-/// A descriptor of a remote repository.
-#[derive(Debug, PartialEq)]
-struct Repository {
-	/// The url of the repository.
-	url: Url,
-	/// If applicable, the branch or tag to be used.
-	reference: Option<String>,
-	/// The name of a package within the repository. Defaults to the repository name.
-	package: String,
-}
-
-impl Repository {
-	/// Parses a url in the form of https://github.com/org/repository?package#tag into its component parts.
-	///
-	/// # Arguments
-	/// * `url` - The url to be parsed.
-	fn parse(url: &str) -> Result<Self, Error> {
-		let url = Url::parse(url)?;
-		let package = url.query();
-		let reference = url.fragment().map(|f| f.to_string());
-
-		let mut url = url.clone();
-		url.set_query(None);
-		url.set_fragment(None);
-
-		let package = match package {
-			Some(b) => b,
-			None => GitHub::name(&url)?,
-		}
-		.to_string();
-
-		Ok(Self { url, reference, package })
-	}
-}
-
-/// Trait for observing status updates.
-pub trait Status {
-	/// Update the observer with the provided `status`.
-	fn update(&self, status: &str);
-}
-
-impl Status for () {
-	// no-op: status updates are ignored
-	fn update(&self, _: &str) {}
-}
-
 /// Attempts to resolve the package manifest from the specified path.
 ///
 /// # Arguments
@@ -859,32 +643,6 @@ fn resolve_manifest(package: &str, path: &Path) -> Result<Option<PathBuf>, Error
 	Ok(manifest.map(|p| p.join("Cargo.toml")))
 }
 
-/// Determines the target triple based on the current platform.
-fn target() -> Result<&'static str, Error> {
-	use std::env::consts::*;
-
-	if OS == "windows" {
-		return Err(Error::UnsupportedPlatform { arch: ARCH, os: OS });
-	}
-
-	match ARCH {
-		"aarch64" => {
-			return match OS {
-				"macos" => Ok("aarch64-apple-darwin"),
-				_ => Ok("aarch64-unknown-linux-gnu"),
-			}
-		},
-		"x86_64" | "x86" => {
-			return match OS {
-				"macos" => Ok("x86_64-apple-darwin"),
-				_ => Ok("x86_64-unknown-linux-gnu"),
-			}
-		},
-		&_ => {},
-	}
-	Err(Error::UnsupportedPlatform { arch: ARCH, os: OS })
-}
-
 #[cfg(test)]
 mod tests {
 	use super::*;
@@ -895,7 +653,14 @@ mod tests {
 
 	mod zombienet {
 		use super::*;
-		use sourcing::tests::Output;
+		use pop_common::sourcing::binary::Status;
+
+		pub(crate) struct Output;
+		impl Status for Output {
+			fn update(&self, status: &str) {
+				println!("{status}")
+			}
+		}
 
 		#[tokio::test]
 		async fn new_with_relay_only_works() -> Result<()> {
@@ -2052,7 +1817,7 @@ node_spawn_timeout = 300
 
 	mod parachain {
 		use super::*;
-		use crate::up::sourcing::GitHub::SourceCodeArchive;
+		use pop_common::sourcing::sourcing::GitHub::SourceCodeArchive;
 		use std::path::PathBuf;
 
 		#[test]
